+++ OnePage/OnePage/Models/Contact.cs
@@ -7,7 +7,7 @@ namespace OnePage.Models
 {
     public class Contact
     {
-        public int id { get; set; }
+        public string id { get; set; }
         public string name { get; set; }
         public string lastName { get; set; }
     }
